Fixed the sign in page with google not registering info on backend
